+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/ImageRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.schedule.dao.impl;
+package com.example.namo2.domain.schedule.dao.repository;
 
 import com.example.namo2.domain.schedule.dao.repository.ImageRepositoryCustom;
 import com.example.namo2.domain.schedule.domain.Schedule;
@@ -9,7 +9,7 @@ import static com.example.namo2.domain.schedule.domain.QImage.image;
 
 import org.springframework.stereotype.Repository;
 
-@Repository
+
 public class ImageRepositoryImpl implements ImageRepositoryCustom {
     private final JPAQueryFactory queryFactory;
     private final EntityManager em;
